requirements.txt
flask
openai
fpdf2
gunicorn
